Разделены функции и описания функций для змейки
